Rework structure and fix typing
