+++ homeassistant_api/__init__.py
@@ -12,7 +12,6 @@ __all__ = (
     "Event",
     "LogbookEntry",
     "WebsocketClient",
-    "AsyncWebsocketClient",
     "AuthInvalid",
     "AuthOk",
     "AuthRequired",
@@ -22,7 +21,6 @@ __all__ = (
     "EventResponse",
 )
 
-from .asyncwebsocket import AsyncWebsocketClient
 from .client import Client
 from .models.domains import Domain, Service
 from .models.entity import Entity, Group
